+++ tests/model-reducer-test.js
@@ -1,13 +1,12 @@
 /* global it, describe */
 
-import reducer from '../src/model-reducer';
-var util = reducer.util;
-var expect = require('chai').expect;
-var sinon = require('sinon');
+import { createReducer, chainReducers } from '../src';
+const expect = require('chai').expect;
 
-var emptyState = {};
-var savedEmptyState = JSON.parse(JSON.stringify(emptyState));
-var initialState1 = {
+const emptyState = {};
+Object.freeze(emptyState);
+const savedEmptyState = JSON.parse(JSON.stringify(emptyState));
+const initialState1 = {
   entities: {
     _meta: {
       foo: {
@@ -26,8 +25,9 @@ var initialState1 = {
     }
   }
 };
-var savedInitialState1 = JSON.parse(JSON.stringify(initialState1));
-var initialState2 = {
+Object.freeze(savedInitialState1);
+const savedInitialState1 = JSON.parse(JSON.stringify(initialState1));
+const initialState2 = {
   entities: {
     _meta: {
       foo: {
@@ -55,87 +55,47 @@ var initialState2 = {
     }
   }
 };
-var savedInitialState2 = JSON.parse(JSON.stringify(initialState2));
+Object.freeze(savedInitialState2);
+const savedInitialState2 = JSON.parse(JSON.stringify(initialState2));
 
 describe('model-reducer', function () {
-  var fooReducer = reducer({
+  const fooReducer = createReducer({
     actionPrefix: 'FOO',
     entityType: 'foo'
   });
 
-  describe('actions', function () {
-    it('clear', function () {
-      var state = util(initialState2).clear('foo').execute();
-      expect(state).to.deep.equal({entities:{_meta:{}}});
-      expect(initialState2).to.deep.equal(savedInitialState2);
-    });
-
-    describe('item operations', function () {
-      it('replace', function () {
-        var state = util(initialState2)
-          .replace('1', 'foo', { test1: 'test1' })
-          .replace('2', 'foo', { test2: 'test2' }, { meta2: 'test2' })
-          .replace('3', 'foo', { test3: 'test3' })
-          .execute();
-        expect(state).to.deep.equal({entities:{_meta:{foo:{
-          '1':{data:{customMetaProp:'foo'}},
-          '2':{data:{customMetaProp:'bar'}},
-          id:{'data':{'meta2':'test2'}}}},'foo':{'1':{'foo':'abc','beep':'boop'},'2':{'abc':'def','ghi':'jkl'},'id':{'test3':'test3'}}}});
-      });
-      it('delete', function () {
-        var spy = sinon.spy();
-        var state = util(initialState2)
-          .iterate('foo', function (id, value, meta) {
-            spy(id, value, meta);
-            if (id === '1') {
-              this.delete(id, 'foo');
-            }
-          })
-          .execute();
-        expect(spy.callCount).to.equal(2);
-        expect(spy.firstCall.args).to.deep.equal([
-          '1',
-          { foo: 'abc', beep: 'boop' },
-          { data: { customMetaProp: 'foo' } }
-        ]);
-        expect(state).to.deep.equal({entities:{_meta:{foo:{'2':{data:{customMetaProp:'bar'}}}},foo:{'2':{abc:'def',ghi:'jkl'}}}});
-        expect(initialState2).to.deep.equal(savedInitialState2);
-      });
-    });
-  });
-
-  describe('join', function () {
-    var origState = {};
-    var reducer1 = function (state, action) {
+  describe('chainReducers', function () {
+    const origState = {};
+    const reducer1 = function (state, action) {
       if (action === 'foo') {
         return 'foo';
       }
       return state;
     };
-    var reducer2 = function (state, action) {
+    const reducer2 = function (state, action) {
       if (action === 'bar') {
         return 'bar';
       }
       return state;
     };
-    var joined = reducer.join([reducer1, reducer2]);
+    const joined = chainReducers([reducer1, reducer2]);
 
     it('should return orig state if no reducers are matched', function () {
-      var state = joined(origState, 'abc');
+      const state = joined(origState, 'abc');
       expect(state).to.equal(origState);
     });
     it('should return new state if a reducer is matched', function () {
-      var state = joined(origState, 'foo');
+      const state = joined(origState, 'foo');
       expect(state).to.equal('foo');
     });
     it('should work with multiple reducers', function () {
-      var state = joined(origState, 'bar');
+      const state = joined(origState, 'bar');
       expect(state).to.equal('bar');
     });
   });
 
   it ('should return provided state for N/A action type', function () {
-    var state = fooReducer(emptyState, {
+    const state = fooReducer(emptyState, {
       type: 'BAR_FETCH_SUCCESS',
       payload: {
         __meta: {
@@ -149,7 +109,7 @@ describe('model-reducer', function () {
 
   describe('DATA', function () {
     it('should add new data attribute', function () {
-      var state = fooReducer(initialState1, {
+      const state = fooReducer(initialState1, {
         type: 'FOO_DATA',
         payload: {
           id: '1',
@@ -158,7 +118,7 @@ describe('model-reducer', function () {
           }
         }
       });
-      expect(initialState1).to.deep.equal(savedInitialState1);
+      expect(initialState2).to.deep.equal(savedInitialState2);
       expect(state).to.deep.equal({
         entities: {
           _meta: {
@@ -182,7 +142,7 @@ describe('model-reducer', function () {
     });
 
     it('should remove existing data attribute', function () {
-      var state = fooReducer(initialState1, {
+      const state = fooReducer(initialState1, {
         type: 'FOO_DATA',
         payload: {
           id: '1',
@@ -215,7 +175,7 @@ describe('model-reducer', function () {
     });
 
     it('should remove all data state', function () {
-      var state = fooReducer(initialState1, {
+      const state = fooReducer(initialState1, {
         type: 'FOO_DATA',
         payload: {
           id: '1',
@@ -242,9 +202,100 @@ describe('model-reducer', function () {
     });
   });
 
+  describe('lifeycle callbacks', function () {
+    it('beforeReduce', function () {
+      const reducer = createReducer({
+        entityType: 'foo',
+        actionPrefix: 'FOO',
+        beforeReduce: function (data, state) {
+          return Object.assign({ foo: 'bar' }, state);
+        }
+      });
+      const state = reducer(emptyState, {
+        type: 'FOO_FETCH_SUCCESS',
+        payload: {
+          result: '1',
+          entities: {
+            foo: {
+              '1': {
+                beep: 'boop'
+              }
+            }
+          }
+        }
+      });
+      delete state.entities._meta.foo['1'].fetched.completedAt;
+      expect(state).to.deep.equal({
+        foo: 'bar',
+        entities: {
+          _meta: {
+            foo: {
+              '1': {
+                fetched: {
+                  entityType: 'foo',
+                  id: '1',
+                  type: 'full'
+                }
+              }
+            }
+          },
+          foo: {
+            '1': {
+              beep: 'boop'
+            }
+          }
+        }
+      });
+    });
+    it('afterReduce', function () {
+      const reducer = createReducer({
+        entityType: 'foo',
+        actionPrefix: 'FOO',
+        afterReduce: function (data, state) {
+          return Object.assign({ foo: 'bar' }, state);
+        }
+      });
+      const state = reducer(emptyState, {
+        type: 'FOO_FETCH_SUCCESS',
+        payload: {
+          result: '1',
+          entities: {
+            foo: {
+              '1': {
+                beep: 'boop'
+              }
+            }
+          }
+        }
+      });
+      delete state.entities._meta.foo['1'].fetched.completedAt;
+      expect(state).to.deep.equal({
+        foo: 'bar',
+        entities: {
+          _meta: {
+            foo: {
+              '1': {
+                fetched: {
+                  entityType: 'foo',
+                  id: '1',
+                  type: 'full'
+                }
+              }
+            }
+          },
+          foo: {
+            '1': {
+              beep: 'boop'
+            }
+          }
+        }
+      });
+    });
+  });
+
   describe('FETCH_SUCCESS', function () {
     it('should handle empty state', function () {
-      var state = fooReducer(emptyState, {
+      const state = fooReducer(emptyState, {
         type: 'FOO_FETCH_SUCCESS',
         payload: {
           result: '1',
@@ -284,7 +335,7 @@ describe('model-reducer', function () {
     });
 
     it('should update an existing model but keep existing custom meta properties', function () {
-      var state = fooReducer(initialState1, {
+      const state = fooReducer(initialState1, {
         type: 'FOO_FETCH_SUCCESS',
         payload: {
           result: '1',
@@ -326,7 +377,7 @@ describe('model-reducer', function () {
     });
 
     it('should update/add normalized entities', function () {
-      var state = fooReducer(initialState1, {
+      const state = fooReducer(initialState1, {
         type: 'FOO_FETCH_SUCCESS',
         payload: {
           result: '1',
@@ -390,7 +441,7 @@ describe('model-reducer', function () {
     });
 
     it('should handle "data" state', function () {
-      var state = fooReducer(emptyState, {
+      const state = fooReducer(emptyState, {
         type: 'FOO_FETCH_SUCCESS',
         payload: {
           result: '1',
@@ -438,7 +489,7 @@ describe('model-reducer', function () {
 
   describe('FETCH_PENDING', function () {
     it('should handle empty state', function () {
-      var state = fooReducer(emptyState, {
+      const state = fooReducer(emptyState, {
         type: 'FOO_FETCH_PENDING',
         payload: { id: '1' }
       });
@@ -458,7 +509,7 @@ describe('model-reducer', function () {
       });
     });
     it('should keep current model details and set new state', function () {
-      var state = fooReducer(initialState1, {
+      const state = fooReducer(initialState1, {
         type: 'FOO_FETCH_PENDING',
         payload: { id: '1' }
       });
@@ -491,7 +542,7 @@ describe('model-reducer', function () {
 
   describe('FETCH_ERROR', function () {
     it('should handle empty state', function () {
-      var state = fooReducer(emptyState, {
+      const state = fooReducer(emptyState, {
         type: 'FOO_FETCH_ERROR',
         payload: {
           id: '1',
@@ -517,7 +568,7 @@ describe('model-reducer', function () {
       });
     });
     it('should clear out current model details and set new state', function () {
-      var state = fooReducer(initialState1, {
+      const state = fooReducer(initialState1, {
         type: 'FOO_FETCH_ERROR',
         payload: {
           id: '1',
@@ -550,7 +601,7 @@ describe('model-reducer', function () {
 
   describe('ACTION_SUCCESS', function () {
     it('should handle empty state', function () {
-      var state = fooReducer(emptyState, {
+      const state = fooReducer(emptyState, {
         type: 'FOO_ACTION_SUCCESS',
         payload: {
           id: '1',
@@ -580,7 +631,7 @@ describe('model-reducer', function () {
       });
     });
     it('should clear out current model details and set new state', function () {
-      var state = fooReducer(initialState1, {
+      const state = fooReducer(initialState1, {
         type: 'FOO_ACTION_SUCCESS',
         payload: {
           id: '1',
@@ -623,7 +674,7 @@ describe('model-reducer', function () {
 
   describe('ACTION_PENDING', function () {
     it('should handle empty state', function () {
-      var state = fooReducer(emptyState, {
+      const state = fooReducer(emptyState, {
         type: 'FOO_ACTION_PENDING',
         payload: {
           id: '1',
@@ -647,7 +698,7 @@ describe('model-reducer', function () {
       });
     });
     it('should work with existing model', function () {
-      var state = fooReducer(initialState1, {
+      const state = fooReducer(initialState1, {
         type: 'FOO_ACTION_PENDING',
         payload: {
           id: '1',
@@ -683,7 +734,7 @@ describe('model-reducer', function () {
 
   describe('ACTION_ERROR', function () {
     it('should handle empty state', function () {
-      var state = fooReducer(emptyState, {
+      const state = fooReducer(emptyState, {
         type: 'FOO_ACTION_ERROR',
         payload: {
           id: '1',
@@ -712,7 +763,7 @@ describe('model-reducer', function () {
       });
     });
     it('should work with existing model', function () {
-      var state = fooReducer(initialState1, {
+      const state = fooReducer(initialState1, {
         type: 'FOO_ACTION_ERROR',
         payload: {
           id: '1',
@@ -754,7 +805,7 @@ describe('model-reducer', function () {
 
   describe('ACTION_CLEAR', function () {
     it('should handle empty state', function () {
-      var state = fooReducer(emptyState, {
+      const state = fooReducer(emptyState, {
         type: 'FOO_ACTION_CLEAR',
         payload: { id: '1' }
       });
@@ -770,7 +821,7 @@ describe('model-reducer', function () {
       });
     });
     it('should work with existing model', function () {
-      var state = fooReducer(initialState1, {
+      const state = fooReducer(initialState1, {
         type: 'FOO_ACTION_CLEAR',
         payload: { id: '1' }
       });
@@ -797,8 +848,8 @@ describe('model-reducer', function () {
     });
 
     it('should actually clear out an action', function () {
-      var state = JSON.parse(JSON.stringify(initialState1));
-      var meta = state.entities._meta.foo['1'];
+      let state = JSON.parse(JSON.stringify(initialState1));
+      const meta = state.entities._meta.foo['1'];
       meta.actionId = 'actionId';
       meta.actionResponse = 'foo';
       meta.actionError = 'bar';
@@ -807,7 +858,7 @@ describe('model-reducer', function () {
         type: 'FOO_ACTION_CLEAR',
         payload: { id: '1' }
       });
-      expect(initialState1).to.deep.equal(savedInitialState1);
+      expect(initialState2).to.deep.equal(savedInitialState2);
       expect(state).to.deep.equal({
         entities: {
           _meta: {